+++ Entidades/BaseEntidade.cs
@@ -1,6 +1,4 @@
-using AutoGestao.Atributes;
 using AutoGestao.Attributes;
-using AutoGestao.Enumerador.Gerais;
 using AutoGestao.Interfaces;
 
 namespace AutoGestao.Entidades
@@ -8,7 +6,7 @@ namespace AutoGestao.Entidades
     [Auditable(AuditCreate = true, AuditUpdate = true, AuditDelete = true)]
     public class BaseEntidade : IAuditable
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [AuditIgnore]
         public DateTime DataCadastro { get; set; }
@@ -17,10 +15,10 @@ namespace AutoGestao.Entidades
         public DateTime DataAlteracao { get; set; }
 
         [AuditIgnore]
-        public int? CriadoPorUsuarioId { get; set; }
+        public long? CriadoPorUsuarioId { get; set; }
 
         [AuditIgnore]
-        public int? AlteradoPorUsuarioId { get; set; }
+        public long? AlteradoPorUsuarioId { get; set; }
 
         // Navigation properties para auditoria
         public virtual Usuario? CriadoPorUsuario { get; set; }
