+++ Services/Interface/IAuditService.cs
@@ -5,21 +5,33 @@ namespace AutoGestao.Services.Interface
 {
     public interface IAuditService
     {
-        Task LogAsync(string entidadeNome, string entidadeId, EnumTipoOperacaoAuditoria tipoOperacao,
-                     object? valoresAntigos = null, object? valoresNovos = null,
-                     string[]? camposAlterados = null, string? mensagemErro = null);
+        Task LogAsync(
+            string entidadeNome,
+            string entidadeId,
+            EnumTipoOperacaoAuditoria tipoOperacao,
+            object? valoresAntigos = null,
+            object? valoresNovos = null,
+            string[]? camposAlterados = null,
+            string? mensagemErro = null);
 
-        Task LogLoginAsync(int usuarioId, string usuarioNome, string usuarioEmail, bool sucesso, string? mensagemErro = null);
+        Task LogLoginAsync(long usuarioId, string usuarioNome, string usuarioEmail, bool sucesso, string? mensagemErro = null);
 
         Task LogHttpRequestAsync(string url, string metodo, bool sucesso, long? duracaoMs = null, string? mensagemErro = null);
 
-        Task<List<AuditLog>> GetLogsAsync(int? usuarioId = null, string? entidade = null,
-                                         EnumTipoOperacaoAuditoria? tipoOperacao = null,
-                                         DateTime? dataInicio = null, DateTime? dataFim = null,
-                                         int skip = 0, int take = 50);
+        Task<List<AuditLog>> GetLogsAsync(
+            long? usuarioId = null,
+            string? entidade = null,
+            EnumTipoOperacaoAuditoria? tipoOperacao = null,
+            DateTime? dataInicio = null,
+            DateTime? dataFim = null,
+            int skip = 0,
+            int take = 50);
 
-        Task<int> GetLogsCountAsync(int? usuarioId = null, string? entidade = null,
-                                   EnumTipoOperacaoAuditoria? tipoOperacao = null,
-                                   DateTime? dataInicio = null, DateTime? dataFim = null);
+        Task<int> GetLogsCountAsync(
+            long? usuarioId = null,
+            string? entidade = null,
+            EnumTipoOperacaoAuditoria? tipoOperacao = null,
+            DateTime? dataInicio = null,
+            DateTime? dataFim = null);
     }
 }
